changed chat.py -> app.py, chat.py to be used to manage chats, allow for multiple boards
