+++ src/gui/fsm.py
@@ -1,0 +1,6 @@
+from PyQt5.QtGui import *
+from PyQt5.QtCore import *
+from PyQt5.QtWidgets import *
+
+class MessageFSM(QStateMachine):
+    def __init__(self,)
