+++ src/gui/mqtt.py
@@ -6,19 +6,25 @@ import json
 import mqtt_link as mqtt
 
 class MQTTNetObject(QObject, mqtt.MQTTLink):
-    message = pyqtSignal(str)
-    emoji = pyqtSignal(int)
+    receive = pyqtSignal(str)
+    emoji = pyqtSignal(list)
     def __init__(self, *args, parent=None, **kwargs):
         super().__init__(parent, *args, **kwargs)
 
+    def __parse__(self, message):
+
+
     def receiveMessage(self, message):
         super().receiveMessage(message)
-        self.new_message.emit(json.dumps(self.last_received))
-        emoji = self.get_Emoji_Tag()
-        if emoji:
-            self.emoji.emit(emoji)
-
         
-    def sendMessage(self, message, receiver):
-        self.addText(message, receiver)
+        emojis = []
+        for msg in message['messages']:
+            receive.emit(msg)
+            for emote in msg['emoji']:
+                emojis.append(emote)
+        emoji.emit(emojis)
+        
+    def sendMessage(self, message):
+        msg = self.__parse__(message)
+        self.addText(msg['data'], msg['receiver'], msg['emojis'])
         self.send()
