+++ test/test_app.py
@@ -2,7 +2,7 @@ from PyQt5.QtGui import *
 from PyQt5.QtCore import *
 from PyQt5.QtWidgets import *
 
-import chat
+import app
 
 # @desc
 # initializes all UI widgets
